test(teams-meetings): split caption ownership coverage
